+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 ## Change log
 ----------------------
 
+Version 1.8
+-------------
+
+ADDED:
+
+- new file extensions for several files in enum FileExtension
+- new github-action workflow for 'Java CI with Gradle'
+- new gradle plugin dependency of 'com.diffplug.spotless:spotless-plugin-gradle' to minor version 6.3.0
+- new method that checks if a given file extension is known from a file object
+
+CHANGED:
+
+- update gradle to new version 7.4
+- update of gradle plugin dependency com.github.ben-manes.versions.gradle.plugin to new minor version 0.42.0
+- update of test dependency test-objects in new major version 6
+- update of test dependency jobj-core in new version 5.3
+
 Version 1.7
 -------------
 
@@ -44,7 +61,7 @@ ADDED:
 - new extension class for execute commands on shell
 - new package io.github.astrapi69.io.annotations created and moved from obsolet project 'de.alpharogroup:jcommons-lang' all related classes to it
 - new package io.github.astrapi69.io.shell created and moved from obsolet project 'de.alpharogroup:jcommons-lang' all related classes to it
-  
+
 CHANGED:
 
 - changed unit test framework from testng to junit5
@@ -73,7 +90,7 @@ Version 1.3
 -------------
 
 ADDED:
- 
+
 - new class Serializer that reflects the SerializedObjectExtensions and throws only decorated RuntimeExceptions
 
 CHANGED:
@@ -88,8 +105,8 @@ Version 1.2
 -------------
 
 ADDED:
- 
-- new method for get an InputStream from a file path 
+
+- new method for get an InputStream from a file path
 - new build system gradle
 
 CHANGED:
@@ -108,12 +125,10 @@ Version 1
 -------------
 
 ADDED:
- 
+
 - this changelog file
 - created PULL_REQUEST_TEMPLATE.md file
 - created CODE_OF_CONDUCT.md file
 - created CONTRIBUTING.md file
 - provide package.html for the javadoc of packages
 - moved classes from obsolet jcommons-lang project to this project
-
-